Splits test files in default scenario into docker & nextcloud
